requirements.txt
asgiref==3.4.1
Django==3.2.11
django-crispy-forms==1.13.0
django-filter==21.1
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
